+++ core/src/main/scala/scalaz/Isomorphism.scala
@@ -641,3 +641,42 @@ trait IsomorphismBitraverse[F[_, _], G[_, _]] extends Bitraverse[F] with Isomorp
   def bitraverseImpl[H[_]: Applicative, A, B, C, D](fab: F[A, B])(f: A => H[C], g: B => H[D]): H[F[C, D]] =
     Applicative[H].map(G.bitraverseImpl(iso.to(fab))(f, g))(iso.from.apply)
 }
+
+trait IsomorphismInvariantApplicative[F[_], G[_]] extends InvariantApplicative[F] with IsomorphismInvariantFunctor[F, G]{
+  implicit def G: InvariantApplicative[G]
+  def xproduct0[Z](f: => Z): F[Z] = iso.from(G.xproduct0(f))
+  def xproduct2[Z, A1, A2](a1: => F[A1], a2: => F[A2])(f: (A1, A2) => Z, g: Z => (A1, A2)): F[Z] = iso.from(G.xproduct2(iso.to(a1), iso.to(a2))(f, g))
+}
+
+trait IsomorphismInvariantAlt[F[_], G[_]] extends InvariantAlt[F] with IsomorphismInvariantApplicative[F, G]{
+  implicit def G: InvariantAlt[G]
+  def xcoproduct2[Z, A1, A2](a1: => F[A1], a2: => F[A2])(f: A1 \/ A2 => Z, g: Z => A1 \/ A2): F[Z] = iso.from(G.xcoproduct2(iso.to(a1), iso.to(a2))(f, g))
+}
+
+trait IsomorphismAlt[F[_], G[_]] extends Alt[F] with IsomorphismApplicative[F, G] with IsomorphismInvariantAlt[F, G]{
+  implicit def G: Alt[G]
+  override def alt[A](a1: =>F[A], a2: =>F[A]): F[A] = iso.from(G.alt(iso.to(a1), iso.to(a2)))
+  override def xproduct0[Z](z: =>Z): F[Z] = super[Alt].xproduct0(z)
+  override def xproduct1[Z, A1](a1: =>F[A1])(f: A1 => Z, g: Z => A1): F[Z] = super[Alt].xproduct1(a1)(f, g)
+  override def xproduct2[Z, A1, A2](a1: =>F[A1], a2: =>F[A2])(f: (A1, A2) => Z, g: Z => (A1, A2)): F[Z] = super[Alt].xproduct2(a1, a2)(f, g)
+  override def xproduct3[Z, A1, A2, A3](a1: =>F[A1], a2: =>F[A2], a3: =>F[A3])(f: (A1, A2, A3) => Z, g: Z => (A1, A2, A3)): F[Z] = super[Alt].xproduct3(a1, a2, a3)(f, g)
+  override def xproduct4[Z, A1, A2, A3, A4](a1: =>F[A1], a2: =>F[A2], a3: =>F[A3], a4: =>F[A4])(f: (A1, A2, A3, A4) => Z, g: Z => (A1, A2, A3, A4)): F[Z] = super[Alt].xproduct4(a1, a2, a3, a4)(f, g)
+  override def xcoproduct1[Z, A1](a1: => F[A1])(f: A1 => Z, g: Z => A1): F[Z] = super[Alt].xcoproduct1(a1)(f, g)
+  override def xcoproduct2[Z, A1, A2](a1: => F[A1], a2: => F[A2])(f: A1 \/ A2 => Z, g: Z => A1 \/ A2): F[Z] = super[Alt].xcoproduct2(a1, a2)(f, g)
+  override def xcoproduct3[Z, A1, A2, A3](a1: => F[A1], a2: => F[A2], a3: => F[A3])(f: A1 \/ (A2 \/ A3) => Z, g: Z => A1 \/ (A2 \/ A3)): F[Z] = super[Alt].xcoproduct3(a1, a2, a3)(f, g)
+  override def xcoproduct4[Z, A1, A2, A3, A4](a1: => F[A1], a2: => F[A2], a3: => F[A3], a4: => F[A4])(f: A1 \/ (A2 \/ (A3 \/ A4)) => Z, g: Z => A1 \/ (A2 \/ (A3 \/ A4))): F[Z] = super[Alt].xcoproduct4(a1, a2, a3, a4)(f, g)
+}
+
+trait IsomorphismDecidable[F[_], G[_]] extends Decidable[F] with IsomorphismDivisible[F, G] with IsomorphismInvariantAlt[F, G]{
+  implicit def G: Decidable[G]
+  def choose2[Z, A1, A2](a1: => F[A1], a2: => F[A2])(f: Z => A1 \/ A2): F[Z] = iso.from(G.choose2(iso.to(a1), iso.to(a2))(f))
+  override def xproduct0[Z](z: =>Z): F[Z] = super[Decidable].xproduct0(z)
+  override def xproduct1[Z, A1](a1: =>F[A1])(f: A1 => Z, g: Z => A1): F[Z] = super[Decidable].xproduct1(a1)(f, g)
+  override def xproduct2[Z, A1, A2](a1: =>F[A1], a2: =>F[A2])(f: (A1, A2) => Z, g: Z => (A1, A2)): F[Z] = super[Decidable].xproduct2(a1, a2)(f, g)
+  override def xproduct3[Z, A1, A2, A3](a1: =>F[A1], a2: =>F[A2], a3: =>F[A3])(f: (A1, A2, A3) => Z, g: Z => (A1, A2, A3)): F[Z] = super[Decidable].xproduct3(a1, a2, a3)(f, g)
+  override def xproduct4[Z, A1, A2, A3, A4](a1: =>F[A1], a2: =>F[A2], a3: =>F[A3], a4: =>F[A4])(f: (A1, A2, A3, A4) => Z, g: Z => (A1, A2, A3, A4)): F[Z] = super[Decidable].xproduct4(a1, a2, a3, a4)(f, g)
+  override def xcoproduct1[Z, A1](a1: => F[A1])(f: A1 => Z, g: Z => A1): F[Z] = super[Decidable].xcoproduct1(a1)(f, g)
+  override def xcoproduct2[Z, A1, A2](a1: => F[A1], a2: => F[A2])(f: A1 \/ A2 => Z, g: Z => A1 \/ A2): F[Z] = super[Decidable].xcoproduct2(a1, a2)(f, g)
+  override def xcoproduct3[Z, A1, A2, A3](a1: => F[A1], a2: => F[A2], a3: => F[A3])(f: A1 \/ (A2 \/ A3) => Z, g: Z => A1 \/ (A2 \/ A3)): F[Z] = super[Decidable].xcoproduct3(a1, a2, a3)(f, g)
+  override def xcoproduct4[Z, A1, A2, A3, A4](a1: => F[A1], a2: => F[A2], a3: => F[A3], a4: => F[A4])(f: A1 \/ (A2 \/ (A3 \/ A4)) => Z, g: Z => A1 \/ (A2 \/ (A3 \/ A4))): F[Z] = super[Decidable].xcoproduct4(a1, a2, a3, a4)(f, g)
+}
